cast cholesky

+++ examples/polybench_hls/cholesky.py
@@ -17,11 +17,11 @@ def cholesky_np(A):
         for j in range(i):
             for k in range(j):
                 A[i, j] = A[i, j] - A[i, k] * A[j, k]
-            A[i, j] = A[i, j] / A[j, j]
+            A[i, j] = int(A[i, j] / A[j, j])
         # i == j
         for k in range(i):
             A[i, i] = A[i, i] - A[i, k] * A[i, k]
-        A[i, i] = A[i, i] * 1.0
+        A[i, i] = A[i, i] * 1
         # A[i, i] = np.sqrt(A[i, i] * 1.0)
